third_party/WebKit/LayoutTests/inspector/tracing/timeline-layout-with-invalidations-expected.txt

Tests the Timeline API instrumentation of layout events with invalidations.

PASS
 

Running: testLocalFrame

Running: testSubframe
